完成登入分流(尚未貼住冊頁面cshtml)

+++ contact/Models/t優惠活動.cs
@@ -18,7 +18,7 @@ namespace contact.Models
         public string f活動名稱 { get; set; }
         public string f活動詳情 { get; set; }
         public string f活動折扣碼 { get; set; }
-        public Nullable<decimal> f活動折扣 { get; set; }
+        public Nullable<double> f活動折扣 { get; set; }
         public string f活動起始日 { get; set; }
         public string f活動結束日 { get; set; }
         public string f活動對象 { get; set; }
